fix circular import of stattype
